+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/backup/container/jwe/JwePasswordProtectedBackupData.kt
@@ -1,8 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved
 
-package com.microsoft.did.sdk.datasource.file.models
+package com.microsoft.did.sdk.datasource.backup.container.jwe
 
 import com.microsoft.did.sdk.crypto.protocols.jose.jwe.JweToken
+import com.microsoft.did.sdk.datasource.backup.content.ProtectedBackupData
 
 class JwePasswordProtectedBackupData internal constructor(
     val jweToken: JweToken,
