+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/backup/container/jwe/JwePasswordProtectionMethod.kt
@@ -1,8 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved
 
-package com.microsoft.did.sdk.datasource.file.models
+package com.microsoft.did.sdk.datasource.backup.container.jwe
 
 import com.microsoft.did.sdk.crypto.protocols.jose.jwe.JweToken
+import com.microsoft.did.sdk.datasource.backup.content.ProtectedBackupData
+import com.microsoft.did.sdk.datasource.backup.container.ProtectionMethod
+import com.microsoft.did.sdk.datasource.backup.content.UnprotectedBackupData
 import com.microsoft.did.sdk.util.controlflow.BadPasswordException
 import com.microsoft.did.sdk.util.controlflow.FailedDecryptException
 import com.nimbusds.jose.EncryptionMethod
@@ -17,7 +20,7 @@ class JwePasswordProtectionMethod(
     val password: String
 ) : ProtectionMethod() {
 
-    override fun encrypt(unprotectedBackupData: UnprotectedBackupData, serializer: Json): ProtectedBackupData {
+    override fun wrap(unprotectedBackupData: UnprotectedBackupData, serializer: Json): ProtectedBackupData {
         val data = serializer.encodeToString(unprotectedBackupData)
         val token = JweToken(data)
         val headers = JWEHeader.Builder(JWEAlgorithm.PBES2_HS512_A256KW, EncryptionMethod.A256CBC_HS512)
@@ -30,7 +33,7 @@ class JwePasswordProtectionMethod(
         return JwePasswordProtectedBackupData(token)
     }
 
-    override fun decrypt(protectedBackupData: ProtectedBackupData, serializer: Json): UnprotectedBackupData {
+    override fun unwrap(protectedBackupData: ProtectedBackupData, serializer: Json): UnprotectedBackupData {
         if (protectedBackupData !is JwePasswordProtectedBackupData) throw FailedDecryptException("Protection paramaters do not match backup contents")
         if (password.isEmpty()) throw BadPasswordException("Password can't be empty")
