+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/backup/content/UnprotectedBackupData.kt
@@ -1,13 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved
 
-package com.microsoft.did.sdk.datasource.file.models
+package com.microsoft.did.sdk.datasource.backup.content
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.modules.SerializersModule
-import kotlinx.serialization.modules.polymorphic
-import kotlinx.serialization.modules.subclass
 
 /**
  * BackupData classes represent the data as found in the backup file.
