+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/backup/content/microsoft2020/Microsoft2020UnprotectedBackupData.kt
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved
 
-package com.microsoft.did.sdk.datasource.file.models.microsoft2020
+package com.microsoft.did.sdk.datasource.backup.content.microsoft2020
 
-import com.microsoft.did.sdk.datasource.file.models.UnprotectedBackupData
+import com.microsoft.did.sdk.datasource.backup.content.UnprotectedBackupData
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
