+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/backup/content/microsoft2020/VcMetadata.kt
@@ -1,4 +1,4 @@
-package com.microsoft.did.sdk.datasource.file.models.microsoft2020
+package com.microsoft.did.sdk.datasource.backup.content.microsoft2020
 
 import com.microsoft.did.sdk.credential.service.models.contracts.display.DisplayContract
 import kotlinx.serialization.SerialName
