+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/backup/content/microsoft2020/WalletMetadata.kt
@@ -1,4 +1,4 @@
-package com.microsoft.did.sdk.datasource.file.models.microsoft2020
+package com.microsoft.did.sdk.datasource.backup.content.microsoft2020
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
