+++ sdk/src/test/java/com/microsoft/did/sdk/datasource/file/Microsoft2020BackupProcessorTest.kt
@@ -3,11 +3,12 @@
 package com.microsoft.did.sdk.datasource.file
 
 import android.util.VerifiableCredentialUtil
-import com.microsoft.did.sdk.datasource.file.models.microsoft2020.Microsoft2020Backup
-import com.microsoft.did.sdk.datasource.file.models.microsoft2020.VcMetadata
-import com.microsoft.did.sdk.datasource.file.models.microsoft2020.WalletMetadata
-import com.microsoft.did.sdk.datasource.file.models.microsoft2020.Microsoft2020UnprotectedBackupData
-import com.microsoft.did.sdk.datasource.file.models.microsoft2020.Microsoft2020BackupProcessor
+import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.Microsoft2020Backup
+import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.VcMetadata
+import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.WalletMetadata
+import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.Microsoft2020UnprotectedBackupData
+import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.Microsoft2020BackupProcessor
+import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.RawIdentifierConverter
 import com.microsoft.did.sdk.util.defaultTestSerializer
 import io.mockk.coVerify
 import io.mockk.verify
