+++ sdk/src/test/java/com/microsoft/did/sdk/datasource/file/Microsoft2020UnprotectedBackupDataTest.kt
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved
 
-package com.microsoft.did.sdk.datasource.file.models
+package com.microsoft.did.sdk.datasource.file
 
 import android.util.VerifiableCredentialUtil
-import com.microsoft.did.sdk.datasource.file.models.microsoft2020.Microsoft2020UnprotectedBackupData
-import com.microsoft.did.sdk.datasource.file.models.microsoft2020.VcMetadata
-import com.microsoft.did.sdk.datasource.file.models.microsoft2020.WalletMetadata
+import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.Microsoft2020UnprotectedBackupData
+import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.VcMetadata
+import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.WalletMetadata
 import com.microsoft.did.sdk.di.defaultTestSerializer
 import org.junit.Test
 import kotlin.test.assertEquals
