+++ sdk/src/test/java/com/microsoft/did/sdk/datasource/file/ProtectedBackupDataFactoryTest.kt
@@ -2,10 +2,11 @@
 
 package com.microsoft.did.sdk.datasource.file
 
-import com.microsoft.did.sdk.datasource.file.models.microsoft2020.VcMetadata
-import com.microsoft.did.sdk.datasource.file.models.microsoft2020.WalletMetadata
+import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.VcMetadata
+import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.WalletMetadata
 import android.util.VerifiableCredentialUtil
-import com.microsoft.did.sdk.datasource.file.models.microsoft2020.Microsoft2020UnprotectedBackupData
+import com.microsoft.did.sdk.datasource.backup.container.jwe.JweProtectedBackupFactory
+import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.Microsoft2020UnprotectedBackupData
 import com.microsoft.did.sdk.util.defaultTestSerializer
 import org.junit.Test
 import java.io.ByteArrayInputStream
